refactor task processing with task boundaries
