+++ scripts/ocr_pdf.py
@@ -2,7 +2,6 @@ import os
 import fitz  # For PDF-håndtering
 import asyncio
 from google.cloud import vision
-from tqdm import tqdm
 import sys
 from pathlib import Path
 from project_config import *
@@ -98,66 +97,3 @@ def pdf_to_images(pdf_path):
         print(f"[ERROR] Failed to convert PDF to images: {e}")
         return []
 
-async def process_image(index, image, ocr_progress):
-    """
-    Utfører OCR på én side, oppdaterer global progress og skriver status til linje 2.
-    """
-    result = await asyncio.to_thread(detect_text, image)
-    ocr_progress[index] = 1  # Merk at siden er ferdig prosessert
-    progress_str = "".join(str(x) for x in ocr_progress)
-    update_progress_line2(progress_str)
-    return result
-
-async def main_async():
-    # Les PDF-sti fra dir.json
-    if not DIR_FILE.exists():
-        print(f"[ERROR] Could not find dir.json at {DIR_FILE}")
-        return ""
-    try:
-        with open(DIR_FILE, "r", encoding="utf-8") as f:
-            dir_data = json.load(f)
-    except Exception as e:
-        print(f"[ERROR] Could not read dir.json: {e}")
-        return ""
-
-    pdf_path = dir_data.get("exam", "").strip()
-    path_obj = Path(pdf_path)
-    if not path_obj.is_absolute():
-        candidate = PDF_DIR / path_obj
-        if candidate.exists():
-            path_obj = candidate
-    if not path_obj.exists():
-        print(f"[ERROR] File does not exist: {path_obj}")
-        return ""
-    pdf_path = str(path_obj)
-
-    print(f"\n[INFO] Selected file: {os.path.basename(pdf_path)}\n")
-
-    images = pdf_to_images(pdf_path)
-    if not images:
-        print("[ERROR] No images could be generated from the PDF. Exiting...")
-        return ""
-
-    # Initialiser en liste for OCR-status: 0 = ikke prosessert, 1 = prosessert
-    ocr_progress = [0] * len(images)
-    update_progress_line2("".join(str(x) for x in ocr_progress))  # Oppdaterer med initial status
-
-    print("[INFO] Starting text extraction from PDF pages...\n")
-
-    tasks = [
-        asyncio.create_task(process_image(i, image, ocr_progress))
-        for i, image in enumerate(images)
-    ]
-    results = await asyncio.gather(*tasks)
-
-    # Lag en enkelt tekststreng med tydelig markering av hver side
-    all_text = ""
-    for idx, page_text in enumerate(results, start=1):
-        all_text += f"\n\n=== PAGE {idx} ===\n\n{page_text}"
-
-    print("\n[INFO] Text extraction complete! Returning collected text.\n")
-    return all_text
-
-
-def main():
-    return asyncio.run(main_async())
